+++ lagom/experimental/exceptions.py
@@ -1,7 +1,0 @@
-from ..exceptions import LagomException
-
-
-class DependencyNotDefined(ValueError, LagomException):
-    """The type must be explicitly defined in the container"""
-
-    pass