+++ tests/experimental/conftest.py
@@ -1,8 +1,0 @@
-import pytest
-
-from lagom.experimental.container import ExplicitContainer
-
-
-@pytest.fixture(scope="function")
-def explicit_container():
-    return ExplicitContainer()